feat(nvim): replace ReplaceWithRegister by mini.operators
"vim-scripts/ReplaceWithRegister" stopped working for text-objects
because `gri` and `gra` were conflicting with native keymappings.

As this plugin is not maintained since 2014, we're trying a more modern
version with "mini.operators".

It's a module of the "mini.nvim" plugin so we're trying out. If we add
more modules later, maybe it will deserve a dedicated `mini.lua` file.
Maybe not.

Note: last commit of "vim-scripts/ReplaceWithRegister"
https://github.com/vim-scripts/ReplaceWithRegister/commit/832efc23111d19591d495dc72286de2fb0b09345